+++ discopy/kleisli/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 """
-DisCoPy's Kleisli modules: monad, channel and multiplicative.
+DisCoPy's Kleisli modules: monad, channel, additive and multiplicative.
 
 .. autosummary::
     :template: module.rst
@@ -9,10 +9,11 @@ DisCoPy's Kleisli modules: monad, channel and multiplicative.
 
     discopy.kleisli.monad
     discopy.kleisli.channel
+    discopy.kleisli.additive
     discopy.kleisli.multiplicative
 """
 
-from discopy.kleisli import monad, channel, multiplicative
+from discopy.kleisli import monad, channel, additive, multiplicative
 from discopy.kleisli.monad import (
     Monad, Maybe, Powerset, Subdistribution, Seed)
 from discopy.kleisli.channel import Channel